[QUIDEM-80] Query recorder
Add a component that can record queries to Quidem files and play them
back.

Here is an example:

  // Step 1. A recorder in RECORD mode creates "foo.iq".
  ConnectionFactory postgres;
  Config config = Config.empty()
    .withFile("foo.iq")
    .withMode(RECORD)
    .withConnectionFactory(postgres);
  Recorder recorder = Recorder.create(config);
  recorder.executeQuery("select count(*) from emp", "empCount",
      result -> {
        assertThat(result.getMetaData().getColumnCount(), is(1));
        assertThat(result.next(), is(true));
        assertThat(result.getInt(1), is(13));
      });
  recorder.close();

  // Step 2. A recorder in PLAY mode reads "foo.iq" created by step 1.
  Config config2 = Config.empty()
    .withFile("foo.iq")
    .withMode(PLAY);
  Recorder recorder2 = Recorder.create(config);
  recorder.executeQuery("select count(*) from emp", "empCount",
      result -> {
        assertThat(result.getMetaData().getColumnCount(), is(1));
        assertThat(result.next(), is(true));
        assertThat(result.getInt(1), is(13));
    });
  recorder2.close();

Here is the `foo.iq` file that is generated in step 1 and read in
step 2:

  !use "postgres"

  # empCount
  select count(*) as c from emp;
  c:int
  ======
  13
  !ok

Add a database ('!use db') to each section.

Improve encoding of strings that are null, empty, or contain
commas or single-quotes. We still cannot encode strings that
contain line endings.

Add a test, based on Postgres, for recorders that verify and
populate their database. It is disabled by default; you
should enable it if you have a local Postgres instance.

Fixes #80