Prepare for [QUIDEM-95] Enable google-java-format
Ahead of #95, add `googleformatter-maven-plugin` to the
POM, disabled. It will be enabled next commit, after all
source files have been reformatted.

Reduce Checkstyle's responsibilities. Since
google-java-format is reformatting the code, less is
needed of checkstyle, but we still need it for semantic
checks, such as checking that files have license headers.

Make some manual fixes to the formatting of a few files, so
that they come out better after automatic reformatting.

Move checkstyle later in Maven lifecycle. If it is at an
earlier phase than google-java-format, it complains about
long lines that google-java-format is just about to fix.